Added flex template creation with metadata support and instructions
